Add env wrappers

+++ sdk-core/assembly/region.ts
@@ -4,9 +4,9 @@ export class Region {
   len: u32;
   capacity: u32;
   constructor(data: Uint8Array) {
-    this.offset = data.dataStart;
-    this.len = data.length;
-    this.capacity = data.length;
+    this.offset = data.dataStart as u32;
+    this.len = data.length as u32;
+    this.capacity = data.length as u32;
   }
   read(): Uint8Array {
     let data = new Uint8Array(this.len);
